Change to external USEARCH/VSEARCH pipeline
We now use https://github.com/markschl/uvsnake, which
contains a lot of code in this repo, but as a reusable unit.
The integration as a Snakemake module was only possible
for SINTAX taxonomy assignment, while the clustering itself
is done in a child process. As a consequence, a clustering
workflow from a sequencing run can (like with the other
pipelines) only run on a single cluster node. This should not
be a problem in most cases.

Other changes:
* Input files have to be provided in a sample list (manifest),
  one per run. The code for dynamically obtaining file lists
  in directories or from patterns was moved to
  https://github.com/markschl/ngs-sample-tab.
* The QC file structure was changed again and is now grouped per
  sequencing run
* FastQC is now run on all files from one run
  (was sample-based). The --group-components arg and the whole
  sample counting is therefore obsolete (per-sample processing
  is done in the uvsnake pipeline internally).